# src/integration-test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: teamssender
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:58754
  description: Generated server url
paths:
  /{municipalityId}/teams/messages:
    post:
      tags:
      - teams-sender-resource
      summary: Send a message in Microsoft Teams
      operationId: sendTeamsMessage
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTeamsMessageRequest"
        required: true
      responses:
        "204":
          description: Message sent successfully
        "400":
          description: Incorrect or malformed request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Recipient not found in Microsoft Graph
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Microsoft Graph rejected the credentials
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Microsoft Graph rejected the request (missing permission or
            activity type not registered)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway when communicating with Microsoft Graph
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Unexpected internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
  /api-docs:
    get:
      tags:
      - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    SendTeamsMessageRequest:
      type: object
      description: Request model for sending a Teams message
      properties:
        messageType:
          type: string
          description: Kind of Teams message to send. Currently only NOTIFICATION
            (bell-icon activity notification) is supported.
          enum:
          - NOTIFICATION
          examples:
          - NOTIFICATION
          minLength: 1
        recipient:
          type: string
          format: email
          description: The recipient's UPN / email address
          examples:
          - first.last@municipaladress.com
          minLength: 1
        message:
          type: string
          description: Notification preview text (Teams enforces a 150 character limit
            on the preview).
          examples:
          - Datacenter A is offline
          maxLength: 150
          minLength: 0
        targetUrl:
          type: string
          description: Optional clickable target URL for the notification. Falls back
            to the service-wide default (teams.default-target-url) if omitted.
          examples:
          - https://status.sundsvall.se/incidents/123
      required:
      - message
      - messageType
      - recipient
  securitySchemes: {}